+++ src/lib/layout/search/SearchResultItem.tsx
@@ -41,6 +41,8 @@ const getIcon = (type: Option<SearchResultType>) => {
       return "validator" as IconKeys;
     case "Pool ID":
       return "pool" as IconKeys;
+    case "NFT Collection Address":
+      return "group" as IconKeys;
     default:
       return "list" as IconKeys;
   }