+++ src/lib/services/nft-collection/sequencer.ts
@@ -1,8 +1,23 @@
 import axios from "axios";
 
-import { zCollectionsByAccountResponseSequencer } from "../types";
-import type { HexAddr } from "lib/types";
+import {
+  zCollectionByCollectionAddressResponseSequencer,
+  zCollectionsByAccountResponseSequencer,
+} from "../types";
+import type { HexAddr, HexAddr32 } from "lib/types";
 import { parseWithError } from "lib/utils";
+
+export const getCollectionByCollectionAddressSequencer = async (
+  endpoint: string,
+  collectionAddress: HexAddr32
+) =>
+  axios
+    .get(
+      `${endpoint}/indexer/nft/v1/collections/${encodeURI(collectionAddress)}`
+    )
+    .then(({ data }) =>
+      parseWithError(zCollectionByCollectionAddressResponseSequencer, data)
+    );
 
 export const getCollectionsByAccountSequencer = async (
   endpoint: string,